Use a Debian Stretch based Vagrant box
This updates the Vagrantfile to use an up-to-date box based on Debian
Stretch (which is the base OS we use in Production).

Also removes the alternates for the libjpeg dev package in `conf/packages`
as this works with our deployment tools but not when fed directly to
`apt-get` via `xargs`.