Vagrant: Changes to permissions and provision script user
Rather than installing things as root and changing permissions
recusively on `/vagrant` run the provisioner as the `vagrant` user and
ensure it owns `/vagrant` before beginning. This prevents permissions
changes from showing up as changes in git as this repo has `filemode`
set to `true`.

We could have made the chmod more discerning, but this seems a simpler
approach.