Add script for importing 2020 attendees
Updated script for 2020 attendees CSV. This is a much simpler version
than the previous import script. This does a simple search for existing
people, but otherwise just creates new people. This means there might be
duplicate people in the system, but only the relevant people are linked
to the 2020 conference, so that shouldn't matter.